Move policy_fixture and clean up conf
Closer to correct discovery, but not yet
