+++ placement/conf/placement.py
@@ -9,6 +9,7 @@
 #    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 #    License for the specific language governing permissions and limitations
 #    under the License.
+from __future__ import absolute_import
 
 from keystoneauth1 import loading as ks_loading
 from oslo_config import cfg
